Refactor file & directory structure
- Refactor file structure
- Remove inaccurate web server functionality
