--- clacks/sinks/display.go
@@ -1,1 +1,0 @@
-package sinks
