--- clacks/sinks/doc.go
@@ -1,6 +1,0 @@
-// Package sinks provides utilities to send messages from the current system.
-//
-//These can be sent to a display, a printer, a directory or a shutter.
-//
-//These sinks could then be used as sources for another system.
-package sinks
