--- clacks/sinks/printer.go
@@ -1,1 +1,0 @@
-package sinks
