+++ encoding/image.go
