+++ encoding/image_test.go
