+++ encoding/message.go
